Fit Vector Autoregressive (VAR) Model Parameters using Lasso Regularization with Lambda Search

Fit Vector Autoregressive (VAR) Model Parameters using Lasso Regularization with Lambda Search

Usage

FitVARLassoSearch(
  Ystd,
  Xstd,
  lambdas,
  crit = "ebic",
  max_iter = 10000L,
  tol = 1e-05
)

Arguments

Ystd

Numeric matrix. Matrix of standardized dependent variables (Y).

Xstd

Numeric matrix. Matrix of standardized predictors (X).

lambdas

Numeric vector. Vector of lambda hyperparameters for Lasso regularization.

crit

Character string. Information criteria to use. Valid values include "aic", "bic", and "ebic".

max_iter

Integer. The maximum number of iterations for the coordinate descent algorithm. Default is 10000.

tol

Numeric. Convergence tolerance. The algorithm stops when the change in coefficients between iterations is below this tolerance. Default is 1e-5.

Value

Matrix of estimated autoregressive and cross-regression coefficients.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

Ystd <- StdMat(vark3p2yx$Y)
Xstd <- StdMat(vark3p2yx$X[, -1])
lambdas <- LambdaSeq(Y = Ystd, X = Xstd, n_lambdas = 100)
FitVARLassoSearch(Ystd = Ystd, Xstd = Xstd, lambdas = lambdas)
#>           [,1]     [,2]     [,3]       [,4]      [,5]      [,6]
#> [1,] 0.3429428 0.000000 0.000000 0.08487819 0.0000000 0.0000000
#> [2,] 0.0000000 0.457026 0.000000 0.00000000 0.2120426 0.0000000
#> [3,] 0.0000000 0.000000 0.588127 0.00000000 0.0000000 0.2737749