Parametric Bootstrap Confidence Intervals

Parametric Bootstrap Confidence Intervals

Usage

PBootCI(x, alpha = 0.05)

Arguments

x

Numeric matrix. Output of PBootVAROLS().

alpha

Numeric. Significance level.

Value

A list with two elements, namely ll for the lower limit and ul for the upper limit.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

pb <- PBootVAROLS(data = vark3p2, p = 2, B = 100)
PBootCI(pb)
#> $ll
#>           [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] 0.4168038  0.31531548 -0.05930300 -0.03906045  0.05011455 -0.02788725
#> [2,] 0.6361473 -0.05137194  0.44189265 -0.07426910 -0.10805638  0.16869095
#> [3,] 0.6875680 -0.09472684 -0.04198385  0.54662891 -0.07589397 -0.06363049
#>             [,7]
#> [1,] -0.06875954
#> [2,] -0.03959759
#> [3,]  0.23345342
#> 
#> $ul
#>          [,1]       [,2]       [,3]       [,4]       [,5]       [,6]       [,7]
#> [1,] 1.169274 0.42688669 0.06208789 0.08461327 0.15852468 0.08855443 0.04958884
#> [2,] 1.372253 0.08227703 0.54031773 0.02835124 0.03126643 0.28345000 0.06175373
#> [3,] 1.609256 0.02905355 0.07021983 0.66552207 0.03769595 0.04855517 0.34699565
#>