Parametric Bootstrap Confidence Intervals

Parametric Bootstrap Confidence Intervals

Usage

PBootCI(x, alpha = 0.05)

Arguments

x

Numeric matrix. Output of PBootVAROLS().

alpha

Numeric. Significance level.

Value

A list with two elements, namely ll for the lower limit and ul for the upper limit.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

pb <- PBootVAROLS(data = VAR, p = 2, B = 100)
PBootCI(pb)
#> $ll
#>           [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] 0.4245776  0.30806794 -0.03928117 -0.04052272  0.05559921 -0.03772574
#> [2,] 0.7113780 -0.04701786  0.41885842 -0.07873768 -0.08885749  0.16187514
#> [3,] 0.8210638 -0.09923716 -0.04496188  0.54850174 -0.07000559 -0.06639735
#>             [,7]
#> [1,] -0.06747421
#> [2,] -0.04124432
#> [3,]  0.24019977
#> 
#> $ul
#>          [,1]       [,2]       [,3]       [,4]       [,5]       [,6]       [,7]
#> [1,] 1.103039 0.41937276 0.07914663 0.07631084 0.17185037 0.08244286 0.05158694
#> [2,] 1.470544 0.07456017 0.53559184 0.04097315 0.02827433 0.28380323 0.07083598
#> [3,] 1.581837 0.02776588 0.08254246 0.64912773 0.05100976 0.04669175 0.35495499
#>