Parametric Bootstrap Confidence Intervals

Parametric Bootstrap Confidence Intervals

Usage

PBootCI(x, alpha = 0.05)

Arguments

x

Numeric matrix. Output of PBootVAROLS().

alpha

Numeric. Significance level.

Value

A list with two elements, namely ll for the lower limit and ul for the upper limit.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

pb <- PBootVAROLS(data = vark3p2, p = 2, B = 100)
PBootCI(pb)
#> $ll
#>           [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] 0.4173527  0.30027800 -0.04835785 -0.04134309  0.06177238 -0.03737552
#> [2,] 0.5787333 -0.04076970  0.42288079 -0.07659620 -0.09225338  0.17319959
#> [3,] 0.7625159 -0.09682382 -0.04622537  0.56145250 -0.07425475 -0.06691990
#>             [,7]
#> [1,] -0.07754469
#> [2,] -0.03325600
#> [3,]  0.22915916
#> 
#> $ul
#>          [,1]       [,2]       [,3]       [,4]       [,5]       [,6]       [,7]
#> [1,] 1.217395 0.41285285 0.06726154 0.09083262 0.17704299 0.07798338 0.05058426
#> [2,] 1.390751 0.07283786 0.54676209 0.04574567 0.01291604 0.29162184 0.08163102
#> [3,] 1.473352 0.02912704 0.07905851 0.65847590 0.05923725 0.04388672 0.35528030
#>