Simulate Data from an Autoregressive Model with Constant Term

This function generates synthetic time series data from an autoregressive (AR) model.

Usage

SimAR(time, burn_in, constant, coef, sd)

Arguments

time

Integer. Number of time points to simulate.

burn_in

Integer. Number of burn-in periods before recording data.

constant

Numeric. The constant term of the AR model.

coef

Numeric vector. Autoregressive coefficients.

sd

Numeric. The standard deviation of the random process noise.

Value

Numeric vector containing the simulated time series data.

Details

The SimAR() function generates synthetic time series data from an autoregressive (AR) model. The generated data follows the AR(p) model, where p is the number of coefficients specified in coef. The generated time series data includes a constant term and autoregressive terms based on the provided coefficients. Random noise, sampled from a normal distribution with mean 0 and standard deviation sd, is added to the time series. A burn-in period is specified to exclude initial data points from the output.

The steps in generating the autoregressive time series with burn-in are as follows:

Set the order of the AR model to p based on the length of coef.

Create a vector data of size time + burn_in to store the generated AR time series data.

Create a vector data of size time + burn_in of random process noise from a normal distribution with mean 0 and standard deviation sd.

Generate the autoregressive time series with burn-in using the formula: $$Y_t = constant + \sum_{i=1}^{p} (coef[i] * Y_{t-i}) + noise_t$$ where \(Y_t\) is the time series data at time \(t\), \(constant\) is the constant term, \(coef[i]\) are the autoregressive coefficients, \(Y_{t-i}\) are the lagged data points up to order p, and \(noise_t\) is the random noise at time \(t\).

Remove the burn-in period from the generated time series data.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
SimAR(time = 10, burn_in = 5, constant = 2, coef = c(0.5, -0.3), sd = 0.1)
#>           [,1]
#>  [1,] 2.441758
#>  [2,] 2.756566
#>  [3,] 2.626824
#>  [4,] 2.890127
#>  [5,] 2.644473
#>  [6,] 2.716173
#>  [7,] 3.022073
#>  [8,] 2.418413
#>  [9,] 2.246827
#> [10,] 2.371225