Generate Random Data for the Variance Vector

This function generates random data for the variance vector given by $$ \boldsymbol{\sigma^{2}} = \exp \left( \boldsymbol{\mu} + \boldsymbol{\varepsilon} \right) \quad \text{with} \quad \boldsymbol{\varepsilon} \sim \mathcal{N} \left( \boldsymbol{0}, \boldsymbol{\Sigma} \right) $$.

SimVariance(n, location, chol_scale)

Arguments

n

Integer. Number of samples to generate.

location

Numeric vector. The constant term \(\boldsymbol{\mu}\).

chol_scale

Numeric matrix. Cholesky decomposition of the covariance matrix \(\boldsymbol{\Sigma}\) for the multivariate normal random error \(\boldsymbol{\varepsilon}\).

Value

Matrix with each row containing the simulated variance vector for each sample.

Details

The SimVariance() function generates random data for the variance vector based on the exponential of a multivariate normal distribution. Given the number of samples n, the constant term \(\boldsymbol{\mu}\) represented by the location vector, and the Cholesky decomposition matrix \(\boldsymbol{\Sigma}\) for the multivariate normal random error \(\boldsymbol{\varepsilon}\), the function simulates \(n\) independent samples of the variance vector \(\boldsymbol{\sigma^{2}}\). Each sample of the variance vector \(\boldsymbol{\sigma^{2}}\) is obtained by calculating the exponential of random variations to the mean vector \(\boldsymbol{\mu}\). The random variations are generated using the Cholesky decomposition of the covariance matrix \(\boldsymbol{\Sigma}\). Finally, the function returns a matrix with each column containing the simulated variance vector for each sample.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
n <- 100
location <- c(0.5, -0.2, 0.1)
chol_scale <- chol(
  matrix(
    data = c(1.0, 0.3, 0.3, 0.3, 1.0, 0.2, 0.3, 0.2, 1.0),
    nrow = 3,
    byrow = TRUE
  )
)
SimVariance(n = n, location = location, chol_scale = chol_scale)
#>              [,1]        [,2]        [,3]
#>   [1,]  0.9689893  0.82351573  1.36961352
#>   [2,]  3.7481137  1.81499146  0.81233741
#>   [3,] 15.1416573  4.63615224  2.14187062
#>   [4,]  2.2227268  1.14225080  1.70260655
#>   [5,]  2.4715825  3.28447845  1.45466449
#>   [6,]  0.8823452  1.22741603  0.33978064
#>   [7,]  4.4290116  0.91583989  2.06574245
#>   [8,]  5.9940375  5.05676091  1.46211617
#>   [9,]  1.1732917  1.28438011  0.70286904
#>  [10,]  9.6204947  2.94312273  2.84982846
#>  [11,]  1.7987375  0.54209102  0.94115054
#>  [12,]  1.5698829  0.38279819  0.35051092
#>  [13,]  1.5524481  2.48718707  2.59730589
#>  [14,]  3.3271043  0.40866984  0.89511041
#>  [15,]  0.5529926  2.12834923  0.19161626
#>  [16,] 23.0120425  1.11074074  0.99113853
#>  [17,]  0.7067468  1.24323116  0.66157195
#>  [18,]  1.6969467  2.33091320  0.14749401
#>  [19,]  1.3262444  0.35209211  0.89993770
#>  [20,]  3.6850778  1.53164445  0.80578387
#>  [21,]  1.3586498  1.54190173  0.70260479
#>  [22,]  5.6186116  1.28155759  0.64461130
#>  [23,]  3.0063061  1.47184767  1.37760380
#>  [24,]  3.6198752  0.30918139  0.83662813
#>  [25,]  0.6397306  0.70761195  0.80655415
#>  [26,]  7.8987802  0.37293136  1.32108985
#>  [27,]  3.6544051  0.64427609  2.41698687
#>  [28,]  2.0519894  0.14178351  1.38447512
#>  [29,]  0.2207307  0.39129401  0.13610495
#>  [30,]  0.5642138  0.53145093  0.31934692
#>  [31,]  1.5185833  3.57974922  1.00781674
#>  [32,]  2.8612588  0.54333536  0.66135707
#>  [33,]  3.7674575  0.09054961  4.06838051
#>  [34,]  2.2218250  0.51083586  1.94062543
#>  [35,]  2.2933200  0.73700075  0.54465019
#>  [36,]  1.7918856  0.30161099  0.06932249
#>  [37,]  0.4447227  1.74582997  0.56249779
#>  [38,]  0.4938529  0.79086626  0.67879059
#>  [39,]  1.0237050  1.38645957  0.74662500
#>  [40,]  6.2746773  2.25461583  0.62515291
#>  [41,]  2.7530984  0.42106984  0.41578020
#>  [42,]  0.6371296  0.20967427  0.24115836
#>  [43,]  0.2971539  1.09988787  0.72414107
#>  [44,]  1.8282720  1.92955459  0.53335829
#>  [45,]  7.6166056  0.66170305  0.46539698
#>  [46,]  2.4400793  0.62412445  3.02784164
#>  [47,]  2.4270720  0.50721011  0.34242874
#>  [48,]  1.1695528  1.28465700  0.16698270
#>  [49,]  0.9605867  0.75296441  0.48456075
#>  [50,]  3.5790448  0.39845628  0.82217769
#>  [51,]  3.6263696  2.48244098  1.71908646
#>  [52,]  0.5720557  1.36698688  1.73826915
#>  [53,]  0.3684204  1.38050515  0.53295428
#>  [54,]  1.4786780  0.82739885  2.80684307
#>  [55,]  0.6882503  1.88318866  2.14925220
#>  [56,]  3.1058871  0.82861316  3.20444606
#>  [57,]  1.6537055  0.26279458  0.33198102
#>  [58,]  1.4852374  0.65504308  0.76281149
#>  [59,]  2.1711740  1.45219299  8.69637685
#>  [60,]  0.5199825  0.56986826  0.62925916
#>  [61,]  0.4396344  0.23920884  0.94626920
#>  [62,]  6.4592149  0.54888735  0.23706992
#>  [63,]  7.1699227  2.21954553  4.25241740
#>  [64,]  2.3144717  6.59468288  0.89369928
#>  [65,]  1.4004988  0.65785183  0.86756368
#>  [66,]  1.2824383  0.48134160  0.44599691
#>  [67,]  0.9637559  0.58966612  2.78500997
#>  [68,]  1.2809781  2.42795847  0.70167002
#>  [69,]  2.2205710  2.22918667  2.09495226
#>  [70,]  1.9420871  0.14321625  0.82682320
#>  [71,]  0.2054686  0.48966991  0.13416580
#>  [72,]  1.9895693  2.16661524  8.79889693
#>  [73,]  6.7143693  1.20153781  8.38030409
#>  [74,]  3.2734592  0.29225407  0.99345483
#>  [75,]  2.8982618  1.01818756  1.47202279
#>  [76,]  0.4658719  0.94561749  1.05704317
#>  [77,]  0.4440750  0.32695520  0.53281105
#>  [78,]  0.9861210  1.72965862  0.04426660
#>  [79,]  4.0912222  0.62052835  2.83022618
#>  [80,]  3.9900348  0.43551633  8.30593118
#>  [81,]  1.6592747  0.54164808  5.91245631
#>  [82,]  1.5086188  0.30815644  2.58260877
#>  [83,]  2.3672157  3.69711942  1.74134710
#>  [84,]  5.0445647  0.15265370  2.31972872
#>  [85,]  8.2364658  0.66209451  0.77825987
#>  [86,]  1.2185998  0.82535415  0.24083607
#>  [87,] 13.2395275 22.50835272  1.59118788
#>  [88,] 14.5205233  0.59229273  1.46233061
#>  [89,]  2.4331951  0.75206653  1.55107045
#>  [90,]  5.1641093  0.93703887  1.04057568
#>  [91,]  0.2813798  0.48116977  0.56364833
#>  [92,]  0.5472978  1.15936672  0.71459134
#>  [93,]  7.6323727  0.76398489 17.27942581
#>  [94,] 10.7007401  0.53097771  0.72178520
#>  [95,]  7.6187583  0.24747822  0.63959292
#>  [96,]  0.7930670  0.81994263  0.24520157
#>  [97,]  0.6575776  0.24327644  1.07303089
#>  [98,]  3.5765165  1.13666503  4.44932599
#>  [99,]  3.7701862  3.18441080  4.73897361
#> [100,]  2.0381412  5.56995669  1.58624582