Generate Random Data for the Variance Vector

This function generates random data for the variance vector given by $$ \boldsymbol{\sigma^{2}} = \exp \left( \boldsymbol{\mu} + \boldsymbol{\varepsilon} \right) \quad \text{with} \quad \boldsymbol{\varepsilon} \sim \mathcal{N} \left( \boldsymbol{0}, \boldsymbol{\Sigma} \right) $$.

Usage

SimVariance(n, location, chol_scale)

Arguments

n

Integer. Number of samples to generate.

location

Numeric vector. The constant term \(\boldsymbol{\mu}\).

chol_scale

Numeric matrix. Cholesky decomposition of the covariance matrix \(\boldsymbol{\Sigma}\) for the multivariate normal random error \(\boldsymbol{\varepsilon}\).

Value

Matrix with each row containing the simulated variance vector for each sample.

Details

The SimVariance() function generates random data for the variance vector based on the exponential of a multivariate normal distribution. Given the number of samples n, the constant term \(\boldsymbol{\mu}\) represented by the location vector, and the Cholesky decomposition matrix \(\boldsymbol{\Sigma}\) for the multivariate normal random error \(\boldsymbol{\varepsilon}\), the function simulates \(n\) independent samples of the variance vector \(\boldsymbol{\sigma^{2}}\). Each sample of the variance vector \(\boldsymbol{\sigma^{2}}\) is obtained by calculating the exponential of random variations to the mean vector \(\boldsymbol{\mu}\). The random variations are generated using the Cholesky decomposition of the covariance matrix \(\boldsymbol{\Sigma}\). Finally, the function returns a matrix with each column containing the simulated variance vector for each sample.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
n <- 100
location <- c(0.5, -0.2, 0.1)
chol_scale <- chol(
  matrix(
    data = c(1.0, 0.3, 0.3, 0.3, 1.0, 0.2, 0.3, 0.2, 1.0),
    nrow = 3,
    byrow = TRUE
  )
)
SimVariance(n = n, location = location, chol_scale = chol_scale)
#>          [,1]      [,2]      [,3]      [,4]      [,5]       [,6]      [,7]
#> [1,] 2.411162 2.0805675 5.8905535 9.7310472 1.3809071 18.0030641 1.5929475
#> [2,] 2.314335 1.1208879 2.6968493 0.8846384 1.4102904  0.3661394 1.7244904
#> [3,] 9.023219 0.6114006 0.8008049 1.0550632 0.3928011  1.1357588 0.9201286
#>          [,8]     [,9]     [,10]    [,11]    [,12]     [,13]     [,14]
#> [1,] 8.518413 1.299636 0.8137269 2.295789 2.515093 0.2684928 5.5019661
#> [2,] 1.589991 4.000121 0.1062624 1.523733 1.132471 0.2453807 1.1965883
#> [3,] 2.327313 2.348330 0.4003448 2.417068 1.195842 0.7037820 0.4491773
#>          [,15]     [,16]    [,17]     [,18]    [,19]     [,20]    [,21]
#> [1,] 0.4851028 2.4719880 1.535395 0.3531955 0.833012 1.1410707 1.029228
#> [2,] 1.0779643 1.1380077 1.878149 0.1935811 1.498519 0.9319504 3.568363
#> [3,] 4.7074100 0.7983882 2.331267 0.9969284 1.108334 0.3705615 4.445571
#>          [,22]     [,23]     [,24]    [,25]     [,26]    [,27]    [,28]
#> [1,] 2.0438983 0.9769686 1.1000988 9.769228 0.2693978 5.343591 2.351850
#> [2,] 0.6807050 0.6660331 0.1147645 1.680859 0.2207756 1.903753 1.315225
#> [3,] 0.8711795 1.4651627 1.3204176 1.885479 0.6792902 1.111869 3.186676
#>           [,29]     [,30]     [,31]      [,32]    [,33]    [,34]    [,35]
#> [1,] 14.0531162 22.985060 0.3200881 13.5982709 1.063177 2.552032 8.129326
#> [2,]  0.7802382  1.353824 0.3412073  3.2404811 1.885319 1.032234 1.217287
#> [3,]  7.9460951  3.258142 4.7166373  0.5526564 2.418725 1.907272 3.890175
#>          [,36]     [,37]     [,38]    [,39]     [,40]     [,41]     [,42]
#> [1,] 4.5444327 3.2448709 0.8096216 6.709578 4.3230346 3.9647864 0.3263333
#> [2,] 0.8097037 1.6442813 2.2699300 0.546311 0.3937564 0.9318016 0.8075775
#> [3,] 4.5858940 0.7151723 0.4498511 2.153108 1.6196838 1.6542133 0.3175371
#>          [,43]     [,44]      [,45]     [,46]     [,47]     [,48]     [,49]
#> [1,] 0.5400685 1.9639888 0.09938411 0.8972535 3.1175333 0.8917105 0.5981621
#> [2,] 0.1306254 3.4226143 0.45571616 2.6908371 1.3675902 2.0322417 0.5163771
#> [3,] 0.9665114 0.6242604 0.90242524 1.5290259 0.4902839 2.5098398 0.6123362
#>          [,50]    [,51]     [,52]     [,53]     [,54]     [,55]     [,56]
#> [1,] 2.2372729 7.254861 2.3017575 0.5497763 0.9667993 5.2272670 1.3983580
#> [2,] 0.7891224 2.112052 2.3953557 0.7171745 0.3224755 5.8356233 0.7972637
#> [3,] 0.4292566 2.900494 0.9261802 1.7984006 0.4948115 0.9337199 3.5051640
#>          [,57]     [,58]     [,59]     [,60]     [,61]     [,62]      [,63]
#> [1,] 2.5280599 2.8879361 1.7346047 2.6935677 1.2287084 0.1653217 19.0727111
#> [2,] 0.1381873 0.6790466 1.2967314 0.4239293 0.3747893 0.4135479  0.3009523
#> [3,] 1.2335584 0.3240610 0.6566801 0.4539400 4.4316892 1.2189898  0.9046446
#>          [,64]     [,65]     [,66]     [,67]     [,68]     [,69]     [,70]
#> [1,] 1.6413482 0.4115715 1.5950478 10.578972 0.9063863 1.3158031 1.1914505
#> [2,] 0.8877375 0.2480168 0.3242563  5.456913 0.7440969 0.2927036 0.5596134
#> [3,] 2.1678281 0.2136371 1.2154075  1.574599 1.5115269 1.5703882 1.5332569
#>          [,71]     [,72]      [,73]     [,74]    [,75]    [,76]     [,77]
#> [1,] 1.2986207 0.5508248 0.57442624 0.5601374 1.431616 1.941229 0.2408948
#> [2,] 0.3129776 0.1505122 0.09352616 0.4522145 0.561807 1.571823 0.3187240
#> [3,] 2.3071841 0.4766438 1.05546087 0.4419167 1.053722 1.615378 0.8616891
#>         [,78]      [,79]     [,80]     [,81]     [,82]     [,83]    [,84]
#> [1,] 1.578118 0.27112489 0.8956977 0.3894451 0.6038797 0.2051387 1.360157
#> [2,] 3.412196 0.05279887 0.5169390 0.2682443 0.2624287 0.1165689 0.982318
#> [3,] 1.899042 0.72518635 0.3887976 1.0979665 2.8214886 0.5644423 2.159974
#>         [,85]     [,86]      [,87]     [,88]     [,89]     [,90]    [,91]
#> [1,] 1.889425 2.7410193 13.0516751 0.2554953 1.5247526 1.0536989 1.033417
#> [2,] 2.371142 0.2691482  0.6872898 1.7824784 0.4282314 1.2563884 6.803731
#> [3,] 2.446926 0.8055412  1.5559922 0.6350662 3.3385226 0.9776397 5.524336
#>          [,92]      [,93]     [,94]     [,95]     [,96]     [,97]     [,98]
#> [1,] 1.5671288 0.57098403 27.791985 1.9258245 0.2747210 1.8139405 2.1668542
#> [2,] 0.9468417 0.03664742  5.252261 1.4377740 0.4400595 0.5500023 7.8896317
#> [3,] 1.4497347 2.30289571  2.975280 0.5224296 0.8567507 0.9581275 0.4644264
#>          [,99]   [,100]
#> [1,] 0.5204153 7.977771
#> [2,] 0.2279232 2.436087
#> [3,] 1.5834490 1.189670